requirements.txt
Flask
opencv-python
numpy
pandas
Pillow
face-recognition
openpyxl
dlib